requirements.txt
streamlit==1.24.0
openai==0.27.8
pandas==2.0.2
ta==0.10.2
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3